bug fix for the merge script

+++ 03_filtering/imp/3_merge/3-0_merge_misc.sh
@@ -6,11 +6,30 @@ cat_or_zstdcat () {
     if [ "${file%.zst}.zst" == "${file}" ] ; then zstdcat $file ; else cat $file ; fi
 }
 
+pgen_to_bed () {
+    local pvarfile=$1
+    local bfile=$2
+    local plink_mem=$3
+    local cores=$4
+    local plink_cmd=$5
+    
+    if [ "${pvarfile%.zst}.zst" == "${pvarfile}" ] ; then
+        vzs="vzs"    
+    else
+        vzs=""
+    fi
+    local pfile=$(echo $zstfile | sed -e "s/.zst$//g" | sed -e "s/.pvar$//g")    
+    
+    plink2 --memory ${plink_mem} --threads ${cores} --pfile "${pfile}" "${vzs}" --out "${bfile}" --${plink_cmd}
+}
+
 prep_files () {
     local zstfile=$1
     local prefix=$2
     local tmp_dir=$3
     local helper_R_script=$4
+    local plink_mem=$5
+    local cores=$6
     
     local pfile=$(echo $zstfile | sed -e "s/.zst$//g" | sed -e "s/.pvar$//g")
     local tmp_pvar="${tmp_dir}/${prefix}.tmp.in.pvar"
@@ -18,9 +37,21 @@ prep_files () {
     local tmp_bed="${tmp_dir}/$(basename $pfile).bed"
     local tmp_fam="${tmp_dir}/$(basename $pfile).fam"
 
-    ln -s ${pfile}.fam ${tmp_fam}    
-    ln -s ${pfile}.bed ${tmp_bed}
+    # generate bed/fam file from pfile when needed, otherwise, place a sym link.
+    if [ ! -f ${tmp_bed} ] ; then
+        pgen_to_bed ${zstfile} ${tmp_bed%.bed} ${plink_mem} ${cores} "make-bed"
+        mv ${tmp_bed%.bed}.log ${tmp_bed%.bed}.bed.log
+    else
+        ln -s ${pfile}.bed ${tmp_bed}
+    fi
     
+    if [ ! -f ${tmp_fam} ] ; then
+        pgen_to_bed ${zstfile} ${tmp_bed%.bed} ${plink_mem} ${cores} "make-just-fam"
+        mv ${tmp_bed%.bed}.log ${tmp_bed%.bed}.fam.log
+    else
+        ln -s ${pfile}.fam ${tmp_fam}
+    fi
+
     cat_or_zstdcat ${zstfile} | grep -v '##' > ${tmp_pvar}
     Rscript ${helper_R_script} ${tmp_pvar} ${prefix} ${pfile}.IDs.tsv
     rm ${tmp_pvar}
